[INJIMOB-1761]: Using webpack to build the library and moving the files into src directory
