+++ js/.gitignore
@@ -3,3 +3,4 @@
 .idea
 /.idea/
 /.idea/
+/dist
