+++ js/ReadMe.md
@@ -7,9 +7,9 @@
 - Run below command to run the example application
   - > npm run example
 - To Modify the Template
-  -  update "id" field with url pointing to svg template over here [VCData.js](example%2FVCData.js)
+  -  update "id" field with url pointing to svg template over here [sample-vc.js](example/sample-vc.js)
 - To Modify the VC Data
-  - update VC in [VCData.js](example%2FVCData.js)
+  - update VC in [sample-vc.js](example/sample-vc.js)
 - Run below command to run the tests
 - > npm run test
 
@@ -20,4 +20,5 @@
   - Extracts the svg template url from the render method
   - Downloads the SVG XML string.
   - Parses the template with Credential Values
+  - Generates the QR code and replaces the `qrCodeImage` placeholder
   - Returns the Parsed svg template
