+++ js/example/example-app.js
@@ -1,5 +1,5 @@
-const {VCRenderer} = require("../VCRenderer.js")
-const {VC} = require('./VCData.js')
+const {VCRenderer} = require("../dist/injivcrenderer.bundle")
+const {VC} = require('./sample-vc.js')
  VCRenderer.renderSVG(VC).then(
     (response) => console.log("SVG Image-->",response)
 );
