+++ js/package.json
@@ -5,18 +5,21 @@
   "author": "MOSIP",
   "license": "MIT",
   "description": "JS library to render VC with SVG template support",
-  "main": "VCRenderer.js",
+  "main": "dist/injivcrenderer.bundle.js",
   "scripts": {
-    "example": "cd example && node example-app.js && cd ..",
+    "example": "npm run build && cd example && node example-app.js && cd ..",
     "test": "jest",
-    "build": "npm test",
+    "build": "webpack",
     "verify": "npm test"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
-    "@babel/preset-env": "^7.25.3",
+    "@babel/preset-env": "^7.25.4",
     "babel-jest": "^29.7.0",
-    "jest": "^29.7.0"
+    "babel-loader": "^9.1.3",
+    "jest": "^29.7.0",
+    "webpack": "^5.94.0",
+    "webpack-cli": "^5.1.4"
   },
   "dependencies": {
     "@mosip/pixelpass": "^0.2.1",
